% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEstBenchmark.R
\name{CompEstBenchmark}
\alias{CompEstBenchmark}
\title{Complexity Estimation and Prediction}
\usage{
CompEstBenchmark(to.model, use = "time")
}
\arguments{
\item{to.model}{A data.frame produced by the CompEst() function, comprised of the following columns: size, time, memory, NlogN_X}

\item{use}{a string indicatif if the function deals "time" or "memory" data}
}
\value{
a list with all the fitted complexity model.
}
\description{
_Benchmark procedure to fit complexity functions to a data.frame of time or memory values
}
