% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEstPlot.R
\name{CompEstPlot}
\alias{CompEstPlot}
\title{Complexity Estimation and Prediction}
\usage{
CompEstPlot(to.plot, element_title = list("", ""), use = "time")
}
\arguments{
\item{to.plot}{a dataset produced by CompEst() function}

\item{element_title}{a string that will be added to the subtitle of the plot}

\item{use}{a string indicatif if the function deals "time" or "memory" data}
}
\value{
a ggplot object
}
\description{
_Plot function for the results of algorithms complexity
}
