% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEstPred.R
\name{CompEstPred}
\alias{CompEstPred}
\title{Complexity Estimation and Prediction}
\usage{
CompEstPred(model.list, benchmark, N, use = "time")
}
\arguments{
\item{model.list}{A list containing the fitted complexity functions, produced by CompEst()}

\item{benchmark}{A vector of LOO errors of complexity functions, produced by CompEst()}

\item{N}{number of rows of the whole dataset, produced by CompEst()}

\item{use}{a string indicatif if the function deals "time" or "memory" data}
}
\value{
a string of the predicted time for the whole dataset
}
\description{
_Prediction function for the computation time of a whole dataset
}
