% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupedSampleFracAtLeastOneSample.R
\name{GroupedSampleFracAtLeastOneSample}
\alias{GroupedSampleFracAtLeastOneSample}
\title{Fraction Sampling without empty output}
\usage{
GroupedSampleFracAtLeastOneSample(d_subset, prop, is.random = TRUE)
}
\arguments{
\item{d_subset}{A dtta.frame from which a small sample is to be returned}

\item{prop}{A number between 0 and 1, being the desired sampling fraction.}

\item{is.random}{a boolean. If TRUE, a random sample is drawn, else it takes the head() of the data}
}
\value{
A random sample from the data, of proportion prop, but always returning at least one observation even if prop is too low.
}
\description{
Sample a random proportion of the data, keeping at least one observation
}
\details{
This function is designed to allow its use with group splitting or do.by methods.
}
